# src/main/java/com/rhb/api/controllers/SubmissionApi.java
/*
 * NOTE: This class is auto generated by the swagger code generator program (3.0.34).
 * https://github.com/swagger-api/swagger-codegen
 * Do not edit the class manually.
 */
package com.rhb.api.controllers;

import com.rhb.api.models.AFAIPSubmissionRequestBody;
import com.rhb.api.models.AFSubmissionResponse;

import com.fasterxml.jackson.databind.ObjectMapper;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.enums.ParameterIn;
import io.swagger.v3.oas.annotations.responses.ApiResponses;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import org.slf4j.Logger;
import org.slf4j.LoggerFactory;
import org.springframework.http.HttpStatus;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.RequestBody;
import org.springframework.web.bind.annotation.RequestHeader;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RequestMethod;

import javax.servlet.http.HttpServletRequest;
import java.io.IOException;
import java.util.Optional;

@javax.annotation.Generated(value = "io.swagger.codegen.v3.generators.java.SpringCodegen", date = "2022-12-20T06:47:20.552442Z[GMT]")
@Validated
public interface SubmissionApi {

    Logger log = LoggerFactory.getLogger(SubmissionApi.class);

    default Optional<ObjectMapper> getObjectMapper() {
        return Optional.empty();
    }

    default Optional<HttpServletRequest> getRequest() {
        return Optional.empty();
    }

    default Optional<String> getAcceptHeader() {
        return getRequest().map(r -> r.getHeader("Accept"));
    }

    @Operation(summary = "Adding Operations", description = "Adding Operations", tags = {})
    @ApiResponses(value = {
            @ApiResponse(responseCode = "200", description = "ok", content = @Content(mediaType = "application/json", schema = @Schema(implementation = AFSubmissionResponse.class))),

            @ApiResponse(responseCode = "500", description = "internal server error", content = @Content(mediaType = "application/json", schema = @Schema(implementation = AFSubmissionResponse.class)))})
    @RequestMapping(value = "/aip/v1/AFApplication/submission",
            produces = {"application/json"},
            consumes = {"application/json"},
            method = RequestMethod.POST)
    default ResponseEntity<AFSubmissionResponse> addingOperations(@Parameter(in = ParameterIn.DEFAULT, description = "AF AIP Submission Request Body", required = true, schema = @Schema()) @RequestBody AFAIPSubmissionRequestBody request, @Parameter(in = ParameterIn.HEADER, description = "Partner Email", required = true, schema = @Schema())
    @RequestHeader(value = "respEmail") String respEmail) throws Exception {
        if (getObjectMapper().isPresent() && getAcceptHeader().isPresent()) {
            if (getAcceptHeader().get().contains("application/json")) {
                try {
                    log.info("Inside Controller interface");
                    return new ResponseEntity<>(getObjectMapper().get().readValue("{\n  \"processID\" : \"processID\",\n  \"processDate\" : \"2000-01-23\",\n  \"customerIDNo\" : \"customerIDNo\",\n  \"customerName\" : \"customerName\"\n}", AFSubmissionResponse.class), HttpStatus.NOT_IMPLEMENTED);
                } catch (IOException e) {
                    log.error("Couldn't serialize response for content type application/json", e);
                    return new ResponseEntity<>(HttpStatus.INTERNAL_SERVER_ERROR);
                }
            }
        } else {
            log.warn("ObjectMapper or HttpServletRequest not configured in default SubmissionApi interface so no example is generated");
        }
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);
    }

    @Operation(summary = "HealthCheck for AF Submission", description = "Health check for AF Submission", tags={  })
    @ApiResponses(value = {
            @ApiResponse(responseCode = "200", description = "ok", content = @Content(mediaType = "application/json", schema = @Schema(implementation = String.class))),
            @ApiResponse(responseCode = "500", description = "internal server error", content = @Content(mediaType = "application/json", schema = @Schema(implementation = String.class)))})
    @RequestMapping(value = "/healthcheck",
            produces = { "application/json" },
            method = RequestMethod.GET)
    default ResponseEntity<String> healthCheck() {
        return new ResponseEntity<>(HttpStatus.OK);
    }

}